+++ src/components/film-details.js
@@ -1,94 +1,14 @@
 import {MONTH_NAMES, EMOJIS} from '../const.js';
-import {getRatingForInsertion, getdurationInHours, getRandomDate} from '../utils/film-utils.js';
+import {getRatingForInsertion, getdurationInHours} from '../utils/film-utils.js';
 import {createElement} from '../utils/utils.js';
+import {generationComments} from '../mock/comment.js';
+import {createCommentsTemplate} from './comments.js';
+
+const COMMENT_COUNT = 4;
+const comments = generationComments(COMMENT_COUNT);
 
 const createGenreTemplate = (genres) => {
   return genres.map((genre) => `<span class="film-details__genre">${genre}</span>`).join(`\n`);
-};
-
-const Comments = [
-  {
-    emoji: `smile`,
-    text: `Interesting setting and a good cast`,
-    author: `Tim Macoveev`,
-    date: getRandomDate(2010, 2019),
-  },
-  {
-    emoji: `sleeping`,
-    text: `Booooooooooring`,
-    author: `John Doe`,
-    date: getRandomDate(2010, 2019),
-  },
-  {
-    emoji: `puke`,
-    text: `Very very old. Meh`,
-    author: `John Doe`,
-    date: getRandomDate(2010, 2019),
-  },
-  {
-    emoji: `angry`,
-    text: `Almost two hours? Seriously?`,
-    author: `John Doe`,
-    date: getRandomDate(2010, 2019),
-  }
-];
-
-const createCommentMarkup = (comment) => {
-  const {emoji, text, author, date} = comment;
-  const dateComment = `${date.getFullYear()}/${date.getMonth()}/${date.getDate()} ${date.getHours()}:${date.getMinutes() > 9 ? date.getMinutes() : `0` + date.getMinutes()}`;
-
-  return (
-    `<li class="film-details__comment">
-      <span class="film-details__comment-emoji">
-        <img src="./images/emoji/${emoji}.png" width="55" height="55" alt="emoji-${emoji}">
-      </span>
-      <div>
-        <p class="film-details__comment-text">${text}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${author}</span>
-          <span class="film-details__comment-day">${dateComment}</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-    </li>`
-  );
-};
-
-const createEmojiMarkup = (emoji) => {
-  return (
-    `<input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-${emoji}" value="${emoji}">
-    <label class="film-details__emoji-label" for="emoji-${emoji}">
-      <img src="./images/emoji/${emoji}.png" width="30" height="30" alt="emoji">
-    </label>`
-  );
-};
-
-const createCommentTemplate = (comment) => {
-  const commentCount = comment.length;
-  const commentMarkup = comment.map((it) => createCommentMarkup(it)).join(`\n`);
-  const emojiListMarkup = EMOJIS.map((it) => createEmojiMarkup(it)).join(`\n`);
-
-  return (
-    `<section class="film-details__comments-wrap">
-      <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentCount}</span></h3>
-
-      <ul class="film-details__comments-list">
-        ${commentMarkup}
-      </ul>
-
-      <div class="film-details__new-comment">
-        <div for="add-emoji" class="film-details__add-emoji-label"></div>
-
-        <label class="film-details__comment-label">
-          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-        </label>
-
-        <div class="film-details__emoji-list">
-          ${emojiListMarkup}
-        </div>
-      </div>
-    </section>`
-  );
 };
 
 const createFilmTableRowlMarkup = (row) => {
@@ -106,7 +26,7 @@ const createFilmDetailsTemplate = (movie) => {
 
   const durationInHours = getdurationInHours(duration);
   const releaseDate = `${date.getDate()} ${MONTH_NAMES[date.getMonth()]} ${date.getFullYear()}`;
-  const commentTemplate = createCommentTemplate(Comments);
+  const commentTemplate = createCommentsTemplate(comments);
 
   const genreTerm = genres.length === 1 ? `Genre` : `Genres`;
   const genreValue = createGenreTemplate(genres);
